CIV-1274 refactor change model name as suggeted in pr
