+++ src/main/routes/routes.ts
@@ -14,11 +14,13 @@ import citizenSevereDisabilityController from './features/response/statementOfMe
 import bankAccountsController from './features/response/statementOfMeans/bankAccounts/bankAccountsController';
 import partnerController from './features/response/statementOfMeans/partner/partnerController';
 import partnerDisabilityController from './features/response/statementOfMeans/partner/partnerDisabilityController';
-import partnerSevereDisabilityController from './features/response/statementOfMeans/partner/partnerSevereDisabilityController';
+import partnerSevereDisabilityController
+  from './features/response/statementOfMeans/partner/partnerSevereDisabilityController';
 import employmentStatusController from './features/response/statementOfMeans/employment/employmentStatusController';
 import residenceController from './features/response/statementOfMeans/residenceController';
 import financialDetailsController from './features/response/financialDetails/financialDetailsController';
-import dependantTeenagersController from './features/response/statementOfMeans/dependants/dependantTeenagersController';
+import betweenSixteenAndNineteenController
+  from './features/response/statementOfMeans/dependants/betweenSixteenAndNineteenController';
 
 export default [
   homeController,
@@ -41,5 +43,5 @@ export default [
   residenceController,
   employmentStatusController,
   financialDetailsController,
-  dependantTeenagersController,
+  betweenSixteenAndNineteenController,
 ];
